Moved js tests from payment container tests to component tests
